% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcpp}
\alias{bcpp}
\title{The Botswana Combination Prevention Project}
\format{A data frame with 10465 observations and 5 variables:
\describe{
    \item{group}{numeric, distinct for each randomised community}
    \item{pair.id}{numeric, distinct for each matched pair}
    \item{treat}{indicates assignment to the intervention arm}
    \item{left}{left bound of interval censored weeks to HIV-infection}
    \item{right}{right bound of interval censored weeks to HIV-infection}
}}
\usage{
bcpp
}
\description{
The Botswana Combination Prevention Project (BCPP) was a pair-matched HIV
prevention CRT to test whether a combination treatment and prevention
intervention could reduce population-level cumulative HIV incidence over 3
years of follow-up. A total of 30 communities were randomized: 15 to the
intervention arm (combination treatment and prevention package) and 15 to
the control arm (enhanced standard of care). The primary study endpoint was
cumulative HIV incidence, measured at scheduled study visits as time to
HIV-infection within a cohort of individuals identified as HIV-negative among
a 20\% random sample of eligible households at baseline. That is, we have an
interval-censored time-to-event outcome for each cohort participant. Since
the primary trial results have not yet been published, this is a simulated
data set, which was generated to mimic the BCPP by applying an agent-based
epidemic model to a dynamic network of simulated sexual partnerships
(Goyal, Blitzstein, and De Gruttola, 2013; Wang et al., 2014).
}
\references{
Goyal, R., Blitzstein, J., and De Gruttola, V. (2013). Simulating Bipartite
Networks to Reflect Uncertainty in Local Network Properties. Harvard
University Biostatistics Working Paper Series.

Wang, R., Goyal, R., Lei, Q., Essex, M., and De~Gruttola, V. (2014). Sample
size considerations in the design of cluster randomized trials of combination
HIV prevention. Clinical Trials 11, 309–318.
}
\keyword{datasets}
