% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{CI_mean}
\alias{CI_mean}
\title{Confidence interval for the additive shift in means for a one- or two-sample problem}
\usage{
CI_mean(x, y = NULL, reps = 1000, side = "both", alpha = 0.05)
}
\arguments{
\item{x}{Vector of observations in group 1}

\item{y}{Vector of observations in group 2}

\item{reps}{Number of replications to approximate distribution (default 1000)}

\item{side}{Type of interval, either 'both', 'upper', or 'lower'. Default is 'both'.}

\item{alpha}{Significance level}
}
\value{
a vector containing the desired limits of the confidence set
}
\description{
Invert the permutation test to get a \eqn{1-\alpha} confidence interval for either the shift parameter in a two-sample problem or the center of the distribution in a one-sample problem
}

