% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\name{compute_irr_ts}
\alias{compute_irr_ts}
\title{Compute the IRR test statistic}
\usage{
compute_irr_ts(ratings)
}
\arguments{
\item{ratings}{matrix of dimension [R, Ns]. Each row corresponds to the ratings given by a single rater; columns correspond to items rated.}
}
\value{
concordance of the ratings, where perfect concordance is 1.0
}
\description{
Compute the test statistic \eqn{\rho_s}. See \href{http://statlab.github.io/permute/api/irr.html}{the Statlab documentation} for details.
}

