% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSP.R
\name{CSP}
\alias{CSP}
\title{Description about the function}
\usage{
CSP(y, x, C = 1000, exact = FALSE)
}
\arguments{
\item{y}{A number.}

\item{x}{A number.}
}
\value{
something
}
\description{
Description about the function
}

