% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/npc.R
\name{fisher}
\alias{fisher}
\alias{liptak}
\alias{tippett}
\title{Combining functions}
\usage{
fisher(p)

liptak(p)

tippett(p)
}
\arguments{
\item{p}{Vector of p-values}
}
\description{
Fisher combining function is \eqn{-2\sum \log(p)}.
}
\section{Functions}{
\itemize{
\item \code{liptak}: \eqn{\sum \Phi^{-1}(1-p)}.

\item \code{tippett}: \eqn{max(1-p)}.
}}

