% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fisher_yates}
\alias{fisher_yates}
\title{Fisher-Yates shuffle}
\usage{
fisher_yates(x)
}
\arguments{
\item{x}{A vector to shuffle}
}
\value{
The permuted data
}
\description{
Also known as Knuth shuffle - a method of taking a random permutation of a list
}

