% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FWEminP.R
\name{FWE.minP_old}
\alias{FWE.minP_old}
\title{Description about the function}
\usage{
FWE.minP_old(P)
}
\arguments{
\item{P}{something}
}
\value{
something
}
\description{
Description about the function
}

