% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{gendata_crt}
\alias{gendata_crt}
\title{Generate data from GLMM}
\usage{
gendata_crt(family = gaussian, nclus, size, theta = 0, sigma, mu,
  rho = 0, sd = 1)
}
\arguments{
\item{family}{a description of the error distribution and link function to
be used in the data generation model. This can be a character string
naming a family function, a family function or the result of a call to a
family function. (See \code{\link[stats]{family}} for details.)
Currently \code{gendata_crt} supports gaussian, binomial, and poisson.}

\item{nclus}{number of clusters}

\item{size}{vector of length 2 specifying range of cluster sizes that will be
drawn from a uniform distribution}

\item{theta}{treatment effect}

\item{sigma}{SD of random cluster effect}

\item{mu}{intercept in GLMM}

\item{rho}{between-cluster correlation (assuming exchangeable covariance); if
non-zero, then there is corrrelation between clusters.}

\item{sd}{SD of residual error (only applies if family = gaussian)}
}
\description{
\code{gendata_crt} generates a dataset from a generalized linear mixed model
}
\details{
This function generates data from a generalized linear mixed model, a GLMM,
with a fixed intercept, fixed treatment effect, and random cluster intercept.
The data will be sorted increasing by cluster then subject. The GLMM is
\deqn{g(E[Y_{ki}]) = \mu + \alpha_k + \theta * X_k}
where \eqn{X_k} is the treatment indicator for the \eqn{k}th cluster and
\eqn{\alpha ~ N(0, \sigma^2)}.
}
\examples{
# generate data from GLMM with bernoulli outcome,
# treatment odds ratio of 1.5, baseline odds in
# control group of 0.3, 10 clusters with sizes
# ranging from 100 to 200.

set.seed(444)
ds <- gendata_crt(family = binomial, nclus = 10, size = c(100, 200),
    theta = log(1.5), sigma = 0.5, mu = log(0.3))
head(ds)
#   unique.id clusid id trt y
# 1       1.1      1  1   1 0
# 2       1.2      1  2   1 1
# 3       1.3      1  3   1 0
# 4       1.4      1  4   1 0
# 5       1.5      1  5   1 0
# 6       1.6      1  6   1 0
}
