% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{gendata_swcrt}
\alias{gendata_swcrt}
\title{Generate data from a stepped wedge cluster randomized trial (SW-CRT)}
\usage{
gendata_swcrt(family = gaussian, nclus, size, nstep, theta = 0, sigma,
  Sigma, mu, beta, nu = 0, rho = 0, sd = 1, redist = "normal")
}
\arguments{
\item{family}{a description of the error distribution and link function to
be used in the data generation model. This can be a character string
naming a family function, a family function or the result of a call to a
family function. (See \code{\link[stats]{family}} for details.)
Currently \code{gendata_swcrt} supports gaussian, binomial, and poisson.}

\item{nclus}{total number of clusters}

\item{size}{if numeric, number of individuals per cluster-period; if vector
(must be length 2), specifies the range of cluster-period sizes, each of
which will be drawn independently from a uniform distribution; if matrix
(must be dim = c(nclus, 2)), the kth row specifies the range of
cluster-period sizes for the kth cluster, similarly drawn from a uniform
distribution.}

\item{nstep}{number of randomization periods not including baseline, i.e.
(nstep + 1) = number of periods.}

\item{theta}{treatment effect}

\item{sigma}{SD of random cluster effect; ignored if \code{Sigma} is specified}

\item{Sigma}{optional user-specified covariance matrix for multivariate
normal random cluster effects}

\item{mu}{intercept in GLMM}

\item{beta}{vector of fixed period effects (length = nstep + 1)}

\item{nu}{SD of random cluster-period effects}

\item{rho}{between-cluster correlation (assuming exchangeable covariance); if
non-zero, then there is corrrelation between clusters.}

\item{sd}{SD of residual error (only applies if family = gaussian)}

\item{redist}{assumed distribution for random cluster effects (currently supports
'normal' and 'lognormal')}
}
\description{
\code{gendata_swcrt} generates a dataset from a SW-CRT based on a GLMM
}
\details{
This function generates data from a SW-CRT using a generalized linear mixed
model, a GLMM, with a fixed intercept, fixed treatment effect, fixed period
effect, random cluster effects, and random cluster-period effects.
The data will be sorted increasing by cluster, then period, then subject.The
GLMM is \deqn{g(E[Y_{kij}]) = \mu + \alpha_k + \beta_j + \eta_{jk} + \theta * X_{jk}}
where \eqn{X_{kj}} is the treatment indicator for the \eqn{k}th cluster in
the \eqn{j}th time period (step).
}
\examples{
# generate data from SW-CRT with continuous outcome,
# treatment effect of 1.5, baseline mean outcome in
# control group of 0, 10 clusters, 5 steps, period effects of 0:5,
# with sizes ranging from 20 to 40.

set.seed(444)
ds <- gendata_swcrt(nclus = 10, size = c(20, 40), nstep = 5, theta = 1.5,
                    sigma = 0.5, mu = 0, beta = 0:5, nu = 0.2, sd = 1)
head(ds)
#   cluster period individual cluster.period.individual cluster.period trt          y
# 1       1      1          1                     1.1.1            1.1   0 -0.5186888
# 2       1      1          2                     1.1.2            1.1   0 -0.3058491
# 3       1      1          3                     1.1.3            1.1   0 -1.0040078
# 4       1      1          4                     1.1.4            1.1   0 -0.6668997
# 5       1      1          5                     1.1.5            1.1   0 -0.5208075
# 6       1      1          6                     1.1.6            1.1   0 -1.7543536
}
