% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getInits}
\alias{getInits}
\title{Get initial values for CI search}
\usage{
getInits(model, trtname, runit, trtunit, strat, data, initmethod, alpha,
  obs1, restrictedDf, restrictedDfType)
}
\description{
\code{getinits} returns initial values for CI search, \code{c(lower, upper)}
}
\details{
If \code{initmethod = "asymp"}, initial bounds are based on asymptotic
approximation (e.g. Wald CI for GLM). If \code{initmethod = "perm"}, initial
bounds are based on the permutation approach used in Garthwaite (1996) with
\eqn{\hat{\theta}\pm \{(t_2 - t_1)/2\}}, where \eqn{t_1} and \eqn{t_2} denote
the second smallest and second largest estimates from the permutation test.

This is a general utility function used within \code{permci}.
}
