% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getInits_cont}
\alias{getInits_cont}
\title{Get initial values for CI search (continuous outcome)}
\usage{
getInits_cont(f, trtname, runit, trtunit, strat, data, alpha, obs1)
}
\description{
\code{getinits} returns initial values for CI search, \code{c(lower, upper)}
}
\details{
Initial bounds are based on the permutation approach used in Garthwaite (1996) with
\eqn{\hat{\theta}\pm \{(t_2 - t_1)/2\}}, where \eqn{t_1} and \eqn{t_2} denote
the second smallest and second largest estimates from the permutation test.

This is a general utility function used within \code{permci_cont}.
}
