% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_sample.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{k-sample test for equality of distributions}
\usage{
goodness_of_fit(x, group, reps = 1000, stat = "ks")
}
\arguments{
\item{x}{Vector of observations}

\item{group}{Vector of group memberships, exchangeable under the null}

\item{reps}{Number of replications to approximate distribution (default 1000)}

\item{stat}{either 'ks' (Kolmogorov-Smirnov, default) or 'ad' (Anderson-Darling)}
}
\value{
A vector of length `reps` containing the permutation distribution
}
\description{
k-sample test for equality of distributions
}

