% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC2samples.R
\name{IC}
\alias{IC}
\title{Description about the function}
\usage{
IC(x1, x2, conf.lev = 0.95, B = 1000, max.delta = 20,
  length.delta = 100)
}
\arguments{
\item{x1}{}

\item{x2}{}
}
\value{
something
}
\description{
Description about the function
}

