% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc.R
\name{inverse_n_weight}
\alias{inverse_n_weight}
\title{Inverse-n weight combining function}
\usage{
inverse_n_weight(p, n)
}
\arguments{
\item{p}{Vector of p-values}

\item{n}{Vector of sample sizes. The \eqn{i}th entry is the
sample size used in the \eqn{i}th test}
}
\value{
The combined test statistic
}
\description{
Compute the test statistic \eqn{-\sum p_s/\sqrt{N_s}}
}

