% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\name{irr_npc_distribution}
\alias{irr_npc_distribution}
\title{Combine the IRR statistics across strata}
\usage{
irr_npc_distribution(perm_distr, size, obs_ts = NULL, pvalues = NULL)
}
\arguments{
\item{perm_distr}{Input matrix of dimension [B, S]
Column s is the permutation distribution of \eqn{rho_s}, for \eqn{s = 1, \dots, S}.}

\item{size}{Vector of S sample sizes. 
Entry s is the number of items \eqn{N_s} in stratum s.}

\item{obs_ts}{Optional: the NPC test statistic.
If NULL, the NPC test statistic is computed using \code{pvalues}.}

\item{pvalues}{Optional: a vector of S IRR p-values for each stratum.
Entry s is the p-value for \eqn{rho_s}, the concordance in stratum s.}
}
\value{
A list containing:
\itemize{
\item{obs_npc: observed value of the test statistic for the input data, or the input value of \code{obs_ts} if it was given as input}
\item{geq: integer number of iterations for which the test statistic was greater than or equal to \code{obs_ts}}
\item{reps: number of permutations}
\item{pvalue: geq/reps}
\item{dist: if \code{keep_dist}, the array of values of the irr test statistic from the \code{reps} iterations. Otherwise, NULL.}
}
}
\description{
Use NPC to combine the IRR statistics across strata.
}
\details{
Simulates the permutation distribution of the combined NPC test
    statistic for S matrices of ratings \code{obs_ts} corresponding to
    S strata. The distribution comes from applying \code{obs_ts}
    to each of the S strata.
    
    If \code{obs_ts} is not null, computes the reference value of the test
    statistic before the first permutation. Otherwise, uses the value
    \code{obs_ts} for comparison.
    
    If \code{keep_dist}, return the distribution of values of the test
    statistic; otherwise, return only the number of permutations
    for which the value of the irr test statistic is at least
    as large as \code{obs_ts}.
}

