% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\name{irr_ts_distribution}
\alias{irr_ts_distribution}
\title{Get the permutation distribution of the IRR test statistic}
\usage{
irr_ts_distribution(ratings, obs_ts = NULL, num_perm = 10000,
  keep_dist = FALSE, seed = NULL)
}
\arguments{
\item{ratings}{matrix of dimension [R, Ns]. 
Each row corresponds to the ratings given by a single rater; 
columns correspond to items rated.}

\item{obs_ts}{if None, \code{obs_ts} is calculated as the value 
of the test statistic for the original data}

\item{num_perm}{integer number of random permutations of the elements of each row of ratings}

\item{keep_dist}{Boolean flag for whether to store and return the array 
of permutation values of the irr test statistic}

\item{seed}{Random seed for random number generator. 
If NULL, the pseudorandom number generator is the instance used by the permutation function.}
}
\value{
A list containing:
\itemize{
\item{obs_ts: observed value of the test statistic for the input data, or the input value of \code{obs_ts} if it was given as input}
\item{geq: integer number of iterations for which the test statistic was greater than or equal to \code{obs_ts}}
\item{num_perm: number of permutations}
\item{pvalue: geq/num_perm}
\item{dist: if \code{keep_dist}, the array of values of the irr test statistic from the \code{num_perm} iterations. Otherwise, NULL.}
}
}
\description{
Simulates the permutation distribution of the irr test statistic for
a matrix of ratings ``ratings``
}
\details{
If \code{obs_ts} is not \code{None}, computes the reference value of the test
statistic before the first permutation. Otherwise, uses the value
\code{obs_ts} for comparison.

If \code{keep_dist}, return the distribution of values of the test statistic;
otherwise, return only the number of permutations for which the value of
the irr test statistic is at least as large as \code{obs_ts}.
}

