% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means.R
\name{one_sample}
\alias{one_sample}
\title{One-sample permutation test for the center of a distribution}
\usage{
one_sample(x, reps = 1000, center = 0)
}
\arguments{
\item{x}{Vector of observations}

\item{reps}{Number of replications to approximate distribution (default 1000)}

\item{center}{Optional; specify null center of distribution (default 0)}
}
\value{
A vector of length `reps` containing the permutation distribution
}
\description{
One-sample permutation test for the center of a distribution
}
\examples{
x <- -5:5 + rnorm(11)
distr <- one_sample(x)
pvalue <- t2p(mean(x), distr, alternative = "two-sided")
pvalue

distr2 <- one_sample(x, center = 5)
pvalue2 <- t2p(mean(x), distr2, alternative = "greater")
pvalue2
}

