% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{permci_glm}
\alias{permci_glm}
\alias{permci_ic_sp}
\title{Randomization CI for regression}
\usage{
permci_glm(formula, trtname, runit, strat = NULL, family = gaussian,
  data, nperm = 999, level = 0.95, quietly = F, ncores = 1)

permci_ic_sp(formula, trtname, runit, strat = NULL, data, nperm = 999,
  level = 0.95, quietly = F, ncores = 1)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted. This argument is passed to the corresponding regression
function, e.g. \code{\link[stats]{glm}} (see Details).}

\item{trtname}{character string specifying the name of randomized treatment
variable in data frame (variable to permute)}

\item{runit}{character string specifying the name of unit of randomization
in data frame}

\item{strat}{character string specifying the name of the variable upon which
randomization was stratified}

\item{family}{a description of the error distribution and link function to
be used in the data generation model. This can be a character string
naming a family function, a family function or the result of a call to a
family function. (See \code{\link[stats]{family}} for details.)}

\item{data}{a data frame containing the variables in the model. This argument
is passed to the corresponding regression function, e.g.
\code{\link[stats]{glm}} (see Details).}

\item{level}{two-sided confidence level (e.g. level = 0.95 for 95\% CI)}

\item{quietly}{logical; if TRUE (and if ncores == 1), status updates will be
printed to Console otherwise, suppress updates.}

\item{ncores}{number of cores to use for computation. If ncores > 1, lower
and upper bound search procedures run in parallel.}
}
\description{
Calculate a randomization confidence interval (CI) for a regression parameter.
}
\details{
These functions are used to calculate randomization confidence intervals (CI)
for a regression parameter. These CIs correspond to inverting randomization
tests by using an offset to test non-zero "null" values. To invert the
randomization test, these functions use a computationally efficient CI
algorithm proposed by
\href{http://doi.org/10.2307/2532852}{Garthwaite (1996)}, which is based on
the Robbins-Monro search process.

Different functions
corrrespond to different regression models:
\itemize{
  \item \code{permci_glm}: randomization CI based on
  \code{\link[stats]{glm}}
  \item \code{permtest_ic_sp}: randomization CI based on
  \code{\link[icenReg]{ic_sp}}
}
To ensure correct specification of the parameters passed to the models above
(e.g. \code{formula} in \code{\link[icenReg]{ic_sp}}), please refer to their
documentation.
}
