% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap.R
\name{permci_survreg_SEQ}
\alias{permci_survreg_SEQ}
\title{SEQ version of permci_survreg (for debugging)}
\usage{
permci_survreg_SEQ(formula, trtname, runit, strat = NULL, data,
  dist = "weibull", nperm = 1000, nburn = 0, level = 0.95, init,
  initmethod = "perm", seed, quietly = F, ...)
}
\description{
SEQ version of permci_survreg (for debugging)
}
