% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{permute}
\alias{permute}
\title{Permute randomized treatment assignments}
\usage{
permute(data, trt.name, runit, strat = NULL)
}
\arguments{
\item{data}{a data frame}

\item{trt.name}{character string specifying the name of randomized treatment
variable in data frame (variable to permute)}

\item{runit}{character string specifying the name of unit of randomization
in data frame}

\item{strat}{character string specifying the name of the variable upon which
randomization was stratified}
}
\description{
\code{permute} returns data with treatment assigments permuted
}
\details{
This function inputs a data frame and outputs the same data frame with the
\code{trt.name} column permuted according to the randomization scheme
provided. This is a general utility function used within test and confidence
interval functions.
}
