% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permute_rows}
\alias{permute_rows}
\title{Permute rows of a matrix}
\usage{
permute_rows(x)
}
\arguments{
\item{x}{A matrix or dataframe, or a list of matrices/dataframes of the same size}
}
\value{
The permuted data. If x is a list, the returned list
will have the same permutation of rows in each matrix/dataframe.
}
\description{
This is a helper function for multivariate permutation tests, when we must
permute multiple variables in lockstep in order to preserve correlations between
them and/or to apply NPC.
}

