% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_sample.R, R/utils.R
\name{permute_within_groups}
\alias{permute_within_groups}
\title{Permute within groups}
\usage{
permute_within_groups(x, group)

permute_within_groups(x, group)
}
\arguments{
\item{x}{vector of exchangeable values}

\item{group}{vector of fixed group assignments}

\item{x}{A vector of treatment indicators}

\item{group}{A vector indicating group membership}
}
\value{
a vector of length(x)

The within-group permuted x
}
\description{
Permute values within groups

Permutation of condition within each group.
This is a helper function for stratified permutation tests.
}

