% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permute_within_rows}
\alias{permute_within_rows}
\title{Permute values within rows of a matrix}
\usage{
permute_within_rows(x)
}
\arguments{
\item{x}{A matrix or dataframe}
}
\value{
The matrix with its rows permuted
}
\description{
Permute values within rows of a matrix
}

