% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.permci}
\alias{plot.permci}
\title{Plot randomization confidence interval}
\usage{
\method{plot}{permci}(x)
}
\arguments{
\item{x}{An object of class "permci" returned from one of the permci functions
(e.g. \code{\link[permuter]{permci_glm}}).}

\item{...}{optional arguments to \code{\link[graphics]{plot}}}
}
\description{
This function takes a permci object and produces plots to monitor convergence
of each randomization-based confidence interval bound. The search values are
indicated by black lines, initial values an x, final update an o, estimated
CI values by red horizontal lines.
}
