% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.permtest}
\alias{plot.permtest}
\title{Plot randomization distribution}
\usage{
\method{plot}{permtest}(x, ...)
}
\arguments{
\item{x}{An object of class "permtest" returned from one of the permtest
functions (e.g. \code{\link[permuter]{permtest_glm}}).}

\item{...}{optional arguments to \code{\link[graphics]{hist}}}
}
\description{
This function takes a permtest object and produces a histogram of the Monte
Carlo randomization distribution of the test statistic. The observed value is
indicated by a red vertical line.
}
