% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/t2p.R
\name{pvalue_distr}
\alias{pvalue_distr}
\title{Computes the p-value for every observation in an empirical distribution}
\usage{
pvalue_distr(distr, alternative = "greater")
}
\arguments{
\item{distr}{The empirical distribution, computed by Monte Carlo}

\item{alternative}{P-value desired: 'greater', 'less', 'two-sided' (may specify more than one)}
}
\description{
Computes the p-value for every observation in an empirical distribution
}

