% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Random_Sample}
\alias{Random_Sample}
\title{Cormen et al. Random_Sample}
\usage{
Random_Sample(x, k)
}
\arguments{
\item{x}{A vector to from which to sample}

\item{k}{Desired sample size}
}
\value{
The selected sample
}
\description{
Recursive method to take a simple random sample from a vector, which does not require sorting.
}

