% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoch_ord.r
\name{stoch.ord}
\alias{stoch.ord}
\title{Description about the function}
\usage{
stoch.ord(y, x, alt = c(-1, 1), B = 1000, seed = 101)
}
\arguments{
\item{y}{A number.}

\item{x}{A number.}

\item{alt}{}

\item{B}{}

\item{seed}{}
}
\value{
something
}
\description{
Description about the function
}

