% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{stratified_two_sample}
\alias{stratified_two_sample}
\title{Stratified two-sample permutation test for equality of means}
\usage{
stratified_two_sample(group, response, stratum, stat = "mean", reps = 1000)
}
\arguments{
\item{group}{Vector of group memberships or treatment conditions}

\item{response}{Vector of measured outcomes, same length as group}

\item{stratum}{Vector of stratum assignments, same length as group}

\item{stat}{The test statistic. Default is 'mean'. See details for other options.}

\item{reps}{Number of replications to approximate distribution (default 1000)}
}
\value{
A vector of length `reps` containing the permutation distribution
}
\description{
Stratified two-sample permutation test for equality of means
}
\details{
\enumerate{
\item If stat == 'mean', the test statistic is (mean(x) - mean(y))
       (equivalently, sum(x), since those are monotonically related), omitting
       NaNs, which therefore can be used to code non-responders
\item If stat == 't', the test statistic is the two-sample t-statistic--
         but the p-value is still estimated by the randomization,
         approximating the permutation distribution.
         The t-statistic is computed using t.test(...,var.equal=TRUE)
\item If stat == 'mean_within_strata', the test statistic is the difference
         in means within each stratum, added across strata.
\item If stat is a function (a callable object), the test statistic is
         that function.  The function should take a permutation of the pooled
         data and compute the test function from it.
 }
}

