% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2p.R
\name{t2p}
\alias{t2p}
\title{Use the empirical distribution of a test statistic to get p-values.}
\usage{
t2p(tst, distr, alternative = c("greater", "less", "two-sided"))
}
\arguments{
\item{distr}{The empirical distribution, computed by Monte Carlo}

\item{alternative}{P-value desired: 'greater', 'less', 'two-sided' (may specify more than one)}

\item{t}{The observed test statistic}
}
\value{
the p-value(s)
}
\description{
Use the empirical distribution of a test statistic to get p-values.
}

