% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/t2p.R
\name{t2p_old}
\alias{t2p_old}
\title{Use the empirical distribution of a test statistic to get p-values.}
\usage{
t2p_old(T)
}
\arguments{
\item{T\t\tA}{vector with the observed test statistic in the first entry and the empirical distribution in the remaining entries}
}
\value{
\t\tsomething
}
\description{
Use the empirical distribution of a test statistic to get p-values.
}

