% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Umbrella.R
\name{umbrella}
\alias{umbrella}
\title{Description about the function}
\usage{
umbrella(x, y, B = 1000, repeated = FALSE, trend = FALSE, alt = NULL)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
something
}
\description{
Description about the function
}

