% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{within_group_mean}
\alias{within_group_mean}
\title{Difference in means within groups}
\usage{
within_group_mean(group, response, stratum, groups, strata)
}
\arguments{
\item{group}{Vector of group memberships or treatment conditions}

\item{response}{Vector of measured outcomes, same length as group}

\item{stratum}{Vector of stratum assignments, same length as group}

\item{groups}{Vector of unique group labels}

\item{strata}{Vector of unique stratum labels}
}
\value{
a vector of differences
}
\description{
Compute difference in mean residual between treated and control within each group
}

