% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CSP.R
\name{CSP}
\alias{CSP}
\title{Description about the function}
\usage{
CSP(y, x, C = 1000, exact = FALSE)
}
\arguments{
\item{y}{A number.}

\item{x}{A number.}

\item{C}{}

\item{exact}{}
}
\value{
something
}
\description{
Description about the function
}
\examples{
CSP(1, 1)
CSP(10, 1)
}

