% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FWEminP.R
\name{fwe_minp}
\alias{fwe_minp}
\title{Permutation version of the Bonferroni-Holm minP step-up procedure}
\usage{
fwe_minp(pvalues)
}
\arguments{
\item{pvalues}{vector of p-values}
}
\value{
something
}
\description{
Permutation version of the Bonferroni-Holm minP step-up procedure
}

