% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/k_sample.R
\name{k_sample}
\alias{k_sample}
\title{k-sample permutation test for equality of distributions}
\usage{
k_sample(x, group, reps = 1000)
}
\arguments{
\item{x}{Vector of observations}

\item{group}{Vector of group memberships}

\item{reps}{Number of replications to approximate distribution (default 1000)}
}
\value{
A vector of length `reps` containing the permutation distribution
}
\description{
k-sample permutation test for equality of distributions
}

