% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/k_sample.R
\name{k_sample}
\alias{k_sample}
\title{k-sample permutation test for equality of distributions}
\usage{
k_sample(x, group, group2 = NULL, reps = 1000, stat = "oneway_anova")
}
\arguments{
\item{x}{Vector of observations}

\item{group}{Vector of group memberships, exchangeable under the null}

\item{group2}{Vector of secondary group memberships (optional)}

\item{reps}{Number of replications to approximate distribution (default 1000)}

\item{stat}{Test statistic: either "oneway_anova" (default) or "twoway_anova"}
}
\value{
A vector of length `reps` containing the permutation distribution
}
\description{
k-sample permutation test for equality of distributions
}

