% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/npc.R
\name{npc}
\alias{npc}
\title{Non-parametric combination of tests}
\usage{
npc(pvalues, distr, combine = "fisher", alternatives = "greater")
}
\arguments{
\item{pvalues}{Vector of partial p-values for tests}

\item{distr}{Matrix or dataframe, columns are approimate null distribution for each partial test}

\item{combine}{Combining function (default is Fisher)}

\item{alternatives}{Optional, vector of alternatives for each test (default is all "greater")

Combining function options are:
* Fisher: $\sum_{i}-2\log(p_i)$
* Liptak: $\sum_{i} \Phi^{-1}(1-p_i)$
* Tippett: $\max_i (1-p_i)$

Alternative options are "greater", "less", or "two-sided". If specified, length of alternatives must
either be 1 or match the length of p.}
}
\value{
A single p-value for the global test
}
\description{
Combines partial p-values from individual hypothesis tests $H_{0i}$ against $H_{1i}$
to test the global null hypothesis
}
\details{
$$\cap_{i} H_{0i}$$

against the alternative

$$ \cup_{i} H_{1i}$$

using a combining function.
}

