% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{permci_glm}
\alias{permci_glm}
\alias{permci_ic_sp}
\alias{permci_survreg}
\title{Randomization CI for regression}
\usage{
permci_glm(formula, trtname, runit, strat = NULL, family = gaussian,
  data, nperm = 1000, level = 0.95, init, quietly = F, ncores = 1)

permci_ic_sp(formula, trtname, runit, strat = NULL, data, nperm = 1000,
  level = 0.95, init, quietly = F, ncores = 1)

permci_survreg(formula, trtname, runit, strat = NULL, data,
  dist = "weibull", nperm = 1000, level = 0.95, init, quietly = F,
  ncores = 1)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted. This argument is passed to the corresponding regression
function, e.g. \code{\link[stats]{glm}} (see Details).}

\item{trtname}{character string specifying the name of randomized treatment
variable in \code{data} (variable to permute)}

\item{runit}{character string specifying the name of unit of randomization
in \code{data}}

\item{strat}{an optional character string specifying the name of the variable
in \code{data} upon which randomization was stratified}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
(See \code{\link[stats]{family}} for details.)}

\item{data}{a data frame containing the variables in the model. This argument
is passed to the corresponding regression function, e.g.
\code{\link[stats]{glm}} (see Details).}

\item{nperm}{number of permutations for each randomization CI bound}

\item{level}{two-sided confidence level (e.g. level = 0.95 for 95\% CI)}

\item{init}{vector of initial values for CI, with lower bound as
first element and upper bound as second. If \code{init} not provided,
initial bounds are based on asymptotic approximations.}

\item{quietly}{logical; if TRUE (and if ncores == 1), status updates will be
printed to Console otherwise, suppress updates.}

\item{ncores}{number of cores to use for computation. If ncores > 1, lower
and upper bound search procedures run in parallel.}

\item{dist}{assumed distribution for y variable. If the argument is a
character string, then it is assumed to name an element from
\code{\link[survival]{survreg.distributions}}. These include "weibull",
"exponential", "gaussian", "logistic","lognormal" and "loglogistic".
Otherwise, it is assumed to be a user defined list conforming to the format
described in \code{\link[survival]{survreg.distributions}}.}
}
\description{
Calculate a randomization confidence interval (CI) for a regression parameter.
}
\details{
These functions are used to calculate randomization confidence intervals (CI)
for a regression parameter. These CIs correspond to inverting randomization
tests by using an offset to test non-zero "null" values. To invert the
randomization test, these functions use a computationally efficient CI
algorithm proposed by
\href{http://doi.org/10.2307/2532852}{Garthwaite (1996)}, which is based on
the Robbins-Monro search process.

Different functions
corrrespond to different regression models:
\itemize{
  \item \code{permci_glm}: randomization CI based on
  \code{\link[stats]{glm}}
  \item \code{permci_ic_sp}: randomization CI based on
  \code{\link[icenReg]{ic_sp}}
  \item \code{permci_survreg}: randomization CI based on
  \code{\link[survival]{survreg}}
}
To ensure correct specification of the parameters passed to the models above
(e.g. \code{formula} in \code{\link[icenReg]{ic_sp}}), please refer to their
documentation.
}
