% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{permtest_glm}
\alias{permtest_glm}
\alias{permtest_ic_sp}
\title{Randomization test for regression}
\usage{
permtest_glm(formula, trtname, runit, strat = NULL, family = gaussian,
  data, nperm = 999, ncores = 1, quietly = F)

permtest_ic_sp(formula, trtname, runit, strat = NULL, data,
  nperm = 999, ncores = 1, quietly = F)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted. This argument is passed to the corresponding regression
function, e.g. \code{\link[stats]{glm}} (see Details).}

\item{trtname}{character string specifying the name of randomized treatment
variable in data frame (variable to permute)}

\item{runit}{character string specifying the name of unit of randomization
in data frame}

\item{strat}{character string specifying the name of the variable upon which
randomization was stratified}

\item{family}{a description of the error distribution and link function to
be used in the data generation model. This can be a character string
naming a family function, a family function or the result of a call to a
family function. (See \code{\link[stats]{family}} for details.)}

\item{data}{a data frame containing the variables in the model. This argument
is passed to the corresponding regression function, e.g.
\code{\link[stats]{glm}} (see Details).}

\item{ncores}{number of cores to use for computation. If ncores > 1, permtest
runs in parallel.}

\item{quietly}{logical; if TRUE (and if ncores == 1), status updates will be
printed to Console otherwise, suppress updates.}
}
\description{
Carry out a randomization test for a treatment effect using a regression
parameter estimate as the test statistic.
}
\details{
These functions are used to carry out randomization tests for a treatment
effect based on a parameter estimate from a regression model.
Different functions corrrespond to different regression models:
\itemize{
  \item \code{permtest_glm}: randomization test based on
  \code{\link[stats]{glm}}
  \item \code{permtest_ic_sp}: randomization test based on
  \code{\link[icenReg]{ic_sp}}
}
To ensure correct specification of the parameters passed to the models above
(e.g. \code{formula} in \code{\link[icenReg]{ic_sp}}), please refer to their
documentation.
}
