% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/k_sample.R
\name{permute_within_groups}
\alias{permute_within_groups}
\title{Permute within groups}
\usage{
permute_within_groups(x, group)
}
\arguments{
\item{x}{vector of responses}

\item{group}{vector of group assignments}
}
\value{
a vector of length(x)
}
\description{
Permute values within groups
}

