% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoch_ord2.R
\name{stoch.ord2}
\alias{stoch.ord2}
\title{Description about the function}
\usage{
stoch.ord2(y, x, z = NULL, alt = c(-1, 1), B = 1000, cat = 0,
  rep = FALSE, seed = 101)
}
\arguments{
\item{y}{A number.}

\item{x}{A number.}

\item{z}{}

\item{alt}{}

\item{B}{}

\item{cat}{}

\item{rep}{}

\item{seed}{}
}
\value{
something
}
\description{
Description about the function
}

