% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/t2p.R
\name{t2p}
\alias{t2p}
\title{Use the empirical distribution of a test statistic to get p-values.}
\usage{
t2p(t, distr, alternative = c("upper", "lower", "two-sided"))
}
\arguments{
\item{t}{The observed test statistic}

\item{distr}{The empirical distribution, computed by Monte Carlo}

\item{alternative}{P-value desired: 'upper', 'lower', 'two-sided' (may specify more than one)}
}
\value{
something
}
\description{
Use the empirical distribution of a test statistic to get p-values.
}

