% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2p.R
\name{t2p_old}
\alias{t2p_old}
\title{Use the empirical distribution of a test statistic to get p-values.}
\usage{
t2p_old(T)
}
\arguments{
\item{T}{an array/vector with the observed test statistic in the first entry and the empirical distribution in the remaining entries}
}
\value{
An array of dimension (B+1) by `dim(T)`. The first element (of the first dimension) contains the p-value
}
\description{
Use the empirical distribution of a test statistic to get p-values.
}

