% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{two_sample}
\alias{two_sample}
\title{Two-sample permutation test for equality of means}
\usage{
two_sample(x, y, reps = 1000)
}
\arguments{
\item{x}{Vector of observations in group 1}

\item{y}{Vector of observations in group 2}

\item{reps}{Number of replications to approximate distribution (default 1000)}
}
\value{
A vector of length `reps` containing the permutation distribution
}
\description{
Two-sample permutation test for equality of means
}

