% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_rm}
\alias{update_rm}
\title{Perform one update using Robbins-Monro search process}
\usage{
update_rm(init, thetahat, t, tstar, alpha, i, m, bound)
}
\arguments{
\item{init}{initial (or most recent) estimate (e.g. L_i or U_i)}

\item{thetahat}{point estimate of theta (parameter of interest) based on
original data}

\item{t}{test statistic for new permutation}

\item{tstar}{test statistic for original permutation}

\item{alpha}{corresponds to 100(1 - alpha)\% CI}

\item{i}{iteration of the search process}

\item{m}{an optional initial magnitude of the steps; if left unspecified,
m defaults to recommended value proposed in Garthwaite and Buckland (1992)}

\item{bound}{"lower" or "upper"; i.e. which confidence bound you want}
}
\description{
\code{update_rm} performs one update using Robbins-Monro search process
}
\details{
This function inputs values necessary to perform one update using the
Robbins-Monro search procedure specific to confidence intervals proposed by
\href{http://doi.org/10.2307/2532852}{Garthwaite (1996)}. Most of the
parameter names are consistent with the notation in
\href{http://doi.org/10.2307/2532852}{Garthwaite (1996)}. This is a general
utility function used within test and confidence interval functions.
}
